kwin/requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.2
aioschedule==0.5.2
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.6.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.0.1
pytz==2023.4
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.26
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.0
yarl==1.9.4
YooMoney==0.1.0
